completed customer crud operations

+++ src/main/java/com/subhashCart/services/impl/CustomerServiceImpl.java
@@ -8,7 +8,9 @@ import com.subhashCart.exceptions.CustomerNotFoundException;
 import com.subhashCart.exceptions.LoginException;
 import com.subhashCart.models.Address;
 import com.subhashCart.models.Customer;
+import com.subhashCart.models.UserSession;
 import com.subhashCart.repositories.CustomerDao;
+import com.subhashCart.repositories.SessionDao;
 import com.subhashCart.services.CustomerService;
 import com.subhashCart.services.LoginLogoutService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,7 +22,8 @@ import java.util.Map;
 import java.util.Optional;
 
 @Service
-public class CustomerServiceImpl implements CustomerService {
+public class CustomerServiceImpl implements CustomerService
+{
 
     @Autowired
     private CustomerDao customerDao;     // ctrl+alt+shift+j -->select same word
@@ -28,6 +31,9 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Autowired
     private LoginLogoutService loginService;
+
+    @Autowired
+    private SessionDao sessionDao;
 
     @Override
     public Customer addCustomer(Customer customer) {
@@ -62,8 +68,6 @@ public class CustomerServiceImpl implements CustomerService {
 
         return opt.get();
     }
-
-
 
 
     // Method to get all customers - only selller or admin can get all customers - check validity of seller token
@@ -108,39 +112,52 @@ public class CustomerServiceImpl implements CustomerService {
         if(opt.isEmpty() && res.isEmpty())
             throw new CustomerNotFoundException("Customer does not exist with given mobile no or email-id");
 
-        Customer existingCustomer = opt.get();
+        Customer existingCustomer = null;
 
-        if(customer.getFirstName() != null) {
-            existingCustomer.setFirstName(customer.getFirstName());
-        }
+        if(opt.isPresent())
+            existingCustomer = opt.get();
+        else
+            existingCustomer = res.get();
 
-        if(customer.getLastName() != null) {
-            existingCustomer.setLastName(customer.getLastName());
-        }
+        UserSession user = sessionDao.findByToken(token).get();
 
-        if(customer.getEmailId() != null) {
-            existingCustomer.setEmailId(customer.getEmailId());
-        }
+        if(existingCustomer.getCustomerId() == user.getUserId()) {
 
-        if(customer.getMobileNo() != null) {
-            existingCustomer.setMobileNo(customer.getMobileNo());
-        }
-
-        if(customer.getPassword() != null) {
-            existingCustomer.setPassword(customer.getPassword());
-        }
-
-        if(customer.getAddress() != null) {
-            for(Map.Entry<String, Address> values : customer.getAddress().entrySet()) {
-                existingCustomer.getAddress().put(values.getKey(), values.getValue());
+            if(customer.getFirstName() != null) {
+                existingCustomer.setFirstName(customer.getFirstName());
             }
+
+            if(customer.getLastName() != null) {
+                existingCustomer.setLastName(customer.getLastName());
+            }
+
+            if(customer.getEmailId() != null) {
+                existingCustomer.setEmailId(customer.getEmailId());
+            }
+
+            if(customer.getMobileNo() != null) {
+                existingCustomer.setMobileNo(customer.getMobileNo());
+            }
+
+            if(customer.getPassword() != null) {
+                existingCustomer.setPassword(customer.getPassword());
+            }
+
+            if(customer.getAddress() != null) {
+                for(Map.Entry<String, Address> values : customer.getAddress().entrySet()) {
+                    existingCustomer.getAddress().put(values.getKey(), values.getValue());
+                }
+            }
+
+            customerDao.save(existingCustomer);
+            return existingCustomer;
+
+        }
+        else {
+            throw new CustomerException("Error in updating. Verification failed.");
         }
 
-        System.out.println(existingCustomer);
 
-        customerDao.save(existingCustomer);
-
-        return existingCustomer;
     }
 
 
@@ -162,18 +179,24 @@ public class CustomerServiceImpl implements CustomerService {
 
         Customer existingCustomer = opt.get();
 
-        existingCustomer.setMobileNo(customerUpdateDTO.getMobileNo());
+        UserSession user = sessionDao.findByToken(token).get();
 
-        customerDao.save(existingCustomer);
+        if(user.getUserId() == existingCustomer.getCustomerId()) {
 
-        return existingCustomer;
+            existingCustomer.setMobileNo(customerUpdateDTO.getMobileNo());
+
+            customerDao.save(existingCustomer);
+
+            return existingCustomer;
+        }
+        else {
+            throw new CustomerException("Error in updating. Verification failed.");
+        }
     }
 
     // Method to update password
-
     @Override
-    public Customer updateCustomerPassword(CustomerDTO customerDTO, String token) {
-
+    public Customer updateCustomerPassword(CustomerDTO customerDTO, String token) throws CustomerNotFoundException {
         if(token.contains("customer") == false) {
             throw new LoginException("Invalid session token for customer");
         }
@@ -187,18 +210,22 @@ public class CustomerServiceImpl implements CustomerService {
 
         Customer existingCustomer = opt.get();
 
-        existingCustomer.setPassword(customerDTO.getPassword());
+        UserSession user = sessionDao.findByToken(token).get();
 
-        return customerDao.save(existingCustomer);
+        if(user.getUserId() == existingCustomer.getCustomerId()) {
+
+            existingCustomer.setPassword(customerDTO.getPassword());
+
+            return customerDao.save(existingCustomer);
+        }
+        else {
+            throw new CustomerException("Error in updating password. Verification failed.");
+        }
+
     }
-
-
-
     // Method to delete a customer by mobile id
-
     @Override
     public SessionDTO deleteCustomer(CustomerDTO customerDTO, String token) throws CustomerNotFoundException {
-
         if(token.contains("customer") == false) {
             throw new LoginException("Invalid session token for customer");
         }
@@ -212,19 +239,79 @@ public class CustomerServiceImpl implements CustomerService {
 
         Customer deletedCustomer = opt.get();
 
-        customerDao.delete(deletedCustomer);
+        UserSession user = sessionDao.findByToken(token).get();
 
         SessionDTO session = new SessionDTO();
 
-        session.setMessage("Account Deleted and Logged out");
+        session.setMessage("");
 
         session.setToken(token);
 
-        return loginService.logoutCustomer(session);
+        if(user.getUserId() == deletedCustomer.getCustomerId()) {
+
+            customerDao.delete(deletedCustomer);
+
+            session.setMessage("Account Deleted and Logged out");
+
+            return loginService.logoutCustomer(session);
+
+        }
+        else {
+            throw new CustomerException("Error in deleting. Verification failed.");
+        }
 
     }
 
+    @Override
+    public Customer updateAddress(Address address, String type, String token) throws CustomerException {
+        if(token.contains("customer") == false) {
+            throw new LoginException("Invalid session token for customer");
+        }
 
+        loginService.checkTokenStatus(token);
 
+        UserSession user = sessionDao.findByToken(token).get();
 
+        Optional<Customer> opt = customerDao.findById(user.getUserId());
+
+        if(opt.isEmpty())
+            throw new CustomerNotFoundException("Customer does not exist");
+
+        Customer existingCustomer = opt.get();
+
+        existingCustomer.getAddress().put(type, address);
+
+        return customerDao.save(existingCustomer);
+    }
+
+    @Override
+    public Customer deleteAddress(String type, String token) throws CustomerException, CustomerNotFoundException {
+        if(token.contains("customer") == false) {
+            throw new LoginException("Invalid session token for customer");
+        }
+
+        loginService.checkTokenStatus(token);
+
+        UserSession user = sessionDao.findByToken(token).get();
+
+        Optional<Customer> opt = customerDao.findById(user.getUserId());
+
+        if(opt.isEmpty())
+            throw new CustomerNotFoundException("Customer does not exist");
+
+        Customer existingCustomer = opt.get();
+
+        if(existingCustomer.getAddress().containsKey(type) == false)
+            throw new CustomerException("Address type does not exist");
+
+        existingCustomer.getAddress().remove(type);
+
+        return customerDao.save(existingCustomer);
+    }
 }
+
+
+
+
+
+
